requirements.txt
--extra-index-url https://download.pytorch.org/whl/nightly/cu124
torch==2.7.0
torchvision==0.22.0

pydantic
gradio
matplotlib
tqdm
opencv-python
scipy
einops
trimesh
tensorboard
transformers
pyglet<2
huggingface-hub[torch]
pillow-heif
pyrender
kapture
kapture-localization
numpy==1.24.4
numpy-quaternion
pycolmap  # for pnp
poselib  # for pnp
viser
tyro
ninja
colorama
pytorch-lightning
splines
diffusers
kornia
open-clip-torch
accelerate
imageio[ffmpeg]
roma
spaces
omegaconf
wandb
evo
open3d

-e ./extern/CUT3R/src/croco/models/curope